+++ src/lib/alert.js
@@ -15,3 +15,17 @@ export const alertError = async (message) => {
     text: message,
   });
 };
+
+export const alertConfirm = async (message) => {
+  const response = await Swal.fire({
+    icon: 'question',
+    title: 'Confirm',
+    text: message,
+    showCancelButton: true,
+    confirmButtonText: 'Yes',
+    cancelButtonText: 'No',
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#3085d6',
+  });
+  return response.isConfirmed;
+};