+++ src/routes/(user)/dashboard/contacts/+page.svelte
@@ -1,6 +1,6 @@
 <script>
-  import { alertError } from "$lib/alert";
-  import { contactList } from "$lib/api/ContactApi";
+  import { alertConfirm, alertError, alertSuccess } from "$lib/alert";
+  import { contactDelete, contactList } from "$lib/api/ContactApi";
   import { onMount } from "svelte";
 
   const token = localStorage.getItem('token')
@@ -19,6 +19,23 @@
     return data
 })
   let contacts = $state([])
+
+  async function handleDeleteContact(id) {
+    if(! await alertConfirm('Are you sure you want to delete this contact ?')){
+        return
+    }
+
+    const response = await contactDelete(token, id)
+    const responseBody = await response.json()
+    console.log(responseBody);
+
+    if (response.status === 200) {
+        await alertSuccess('Success delete contact')
+        await fetchContacts()
+    } else {
+        await alertError('Failed delete contact')
+    }
+  }
 
   async function handlePageChange(value) {
     search.page = value
@@ -198,7 +215,7 @@
                         <a href="/dashboard/contacts/{contact.id}/edit" class="px-4 py-2 bg-gradient text-white rounded-lg hover:opacity-90 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 focus:ring-offset-gray-800 transition-all duration-200 font-medium shadow-md flex items-center">
                             <i class="fas fa-edit mr-2"></i> Edit
                         </a>
-                        <button class="px-4 py-2 bg-gradient-to-r from-red-600 to-red-500 text-white rounded-lg hover:opacity-90 focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2 focus:ring-offset-gray-800 transition-all duration-200 font-medium shadow-md flex items-center">
+                        <button onclick={() => handleDeleteContact(contact.id)} class="px-4 py-2 bg-gradient-to-r from-red-600 to-red-500 text-white rounded-lg hover:opacity-90 focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2 focus:ring-offset-gray-800 transition-all duration-200 font-medium shadow-md flex items-center">
                             <i class="fas fa-trash-alt mr-2"></i> Delete
                         </button>
                     </div>